Reorganized testing logic
